+++ ui/mods/com.pa.quitch.ai-chat/live_game/communication.js
@@ -647,6 +647,27 @@ if (!aiCommunicationsLoaded) {
           _.delay(communicateLandingLocation, 10000); // delay to allow AI to spawn
         }
       });
+
+      handlers.kills = function (payload) {
+        console.log("Kill handler called");
+        console.log(payload);
+        var defeated = payload[0];
+        var killer = payload[1];
+        var killerIsAI = players[killer.index].ai === 1;
+        var defeatedIsAIAlly =
+          players[defeated.index].ai === 1 &&
+          players[defeated.index].stateToPlayer === allyState;
+
+        if (killerIsAI) {
+          console.log("Kill message sent");
+          sendMessage("global", killer.name, "kill");
+        }
+
+        if (defeatedIsAIAlly) {
+          console.log("Defeat message sent");
+          sendMessage("team", defeated.name, "defeat");
+        }
+      };
     } catch (e) {
       console.error(e);
       console.error(JSON.stringify(e));