+++ ui/mods/com.pa.quitch.ai-chat/live_game_unit_alert/kills.js
@@ -1,0 +1,23 @@
+var aiKillsLoaded;
+
+if (!aiKillsLoaded) {
+  aiKillsLoaded = true;
+
+  function aiKills() {
+    try {
+      model.defeatedArmyAlerts.subscribe(function (alerts) {
+        console.log("New defeat");
+        alerts.forEach(function (alert) {
+          var defeated = alert.defeated;
+          var killer = alert.killer;
+          console.log(defeated, killer);
+          api.Panel.message(api.Panel.parentId, "kills", [defeated, killer]);
+        });
+      });
+    } catch (e) {
+      console.error(e);
+      console.error(JSON.stringify(e));
+    }
+  }
+  aiKills();
+}